+++ src/main/java/roomescape/service/ThemeService.java
@@ -3,9 +3,9 @@ package roomescape.service;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.domain.ReservationRepository;
 import roomescape.domain.Theme;
-import roomescape.domain.ThemeRepository;
+import roomescape.persistence.ReservationRepository;
+import roomescape.persistence.ThemeRepository;
 import roomescape.service.request.ThemeRequest;
 import roomescape.service.response.ThemeResponse;
