+++ src/test/java/roomescape/persistence/ReservationRepositoryTest.java
@@ -18,7 +18,6 @@ import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 import roomescape.domain.Name;
 import roomescape.domain.Reservation;
-import roomescape.domain.ReservationRepository;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
 
@@ -38,7 +37,7 @@ class ReservationRepositoryTest {
 
     @BeforeEach
     void setUp() {
-        reservationRepository = new JdbcReservationRepository(jdbcTemplate);
+        reservationRepository = new ReservationRepository(jdbcTemplate);
     }
 
     @Test
